requirements.txt
apache-airflow-providers-snowflake
loguru
pandas
snowflake-connector-python
snowflake-connector-python[pandas]
snowflake-sqlalchemy